+++ src/shwetado/evaluator/Operation.java
@@ -4,40 +4,40 @@ import java.util.HashMap;
 import java.util.Map;
 
 interface Operator {
-    void operate(int operand1, int operand2);
+    void operate(double operand1, double operand2);
 }
 
 public class Operation {
-    int result;
-    public int evaluateResult(String operator,int num1,int num2) throws Exception {
+    double result;
+    public double evaluateResult(String operator,double num1,double num2) throws Exception {
         Map<String, Operator> methodMap = new HashMap<String, Operator>();
         methodMap.put("+", new Operator() {
-            public void operate(int operand1, int operand2) {
+            public void operate(double operand1, double operand2) {
                 result = operand1 + operand2;
             }
         });
 
         methodMap.put("-", new Operator() {
-            public void operate(int operand1, int operand2) {
+            public void operate(double operand1, double operand2) {
                 result = operand1 - operand2;
             }
         });
 
         methodMap.put("*", new Operator() {
-            public void operate(int operand1, int operand2) {
+            public void operate(double operand1, double operand2) {
                 result = operand1 * operand2;
             }
         });
 
         methodMap.put("/", new Operator() {
-            public void operate(int operand1, int operand2) {
+            public void operate(double operand1, double operand2) {
                 result = operand1 / operand2;
             }
         });
 
         methodMap.put("^", new Operator() {
-            public void operate(int operand1, int operand2) {
-                result = (int) Math.pow(operand1, operand2);
+            public void operate(double operand1, double operand2) {
+                result = Math.pow((double)operand1, (double)operand2);
             }
         });
